# ttms-backend/internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "ttms影院管理系统",
        "title": "ttms",
        "contact": {},
        "license": {},
        "version": "1.0"
    },
    "host": "127.0.0.1",
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/email/send": {
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "email"
                ],
                "summary": "发邮件",
                "parameters": [
                    {
                        "type": "string",
                        "name": "Email",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "1001:参数有误 1003:系统错误 2001:鉴权失败 20001:用户已存在 30002:发送次数过多",
                        "schema": {
                            "$ref": "#/definitions/common.State"
                        }
                    }
                }
            }
        },
        "/api/v1/file/send": {
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "file"
                ],
                "summary": "上传文件",
                "parameters": [
                    {
                        "type": "string",
                        "description": "x_token 用户令牌",
                        "name": "x_token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "文件",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "1001:参数有误 1003:系统错误 2001:鉴权失败 ",
                        "schema": {
                            "$ref": "#/definitions/common.State"
                        }
                    }
                }
            }
        },
...
        "/api/v1/movie/create": {
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movie"
                ],
                "summary": "创建电影",
                "parameters": [
                    {
                        "type": "string",
                        "description": "x_token 用户令牌",
                        "name": "x_token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "minItems": 1,
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "description": "演员",
                        "name": "actors",
                        "in": "query",
                        "required": true
                    },
                    {
                        "maxLength": 20,
                        "minLength": 1,
                        "type": "string",
                        "description": "地区",
                        "name": "area",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "avatar",
                        "in": "query"
                    },
                    {
                        "minLength": 1,
                        "type": "string",
                        "name": "content",
                        "in": "query",
                        "required": true
                    },
                    {
                        "maxLength": 5,
                        "type": "string",
                        "description": "导演",
                        "name": "director",
                        "in": "query",
                        "required": true
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "时长(分钟)",
                        "name": "duration",
                        "in": "query",
                        "required": true
                    },
                    {
                        "maxLength": 20,
                        "minLength": 1,
                        "type": "string",
                        "description": "电影名称",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "上映时间戳",
                        "name": "show_time",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "description": "电影标签",
                        "name": "tags",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "1001:参数有误 1003:系统错误 2001:鉴权失败 20001:用户已存在 30002:发送次数过多",
                        "schema": {
                            "$ref": "#/definitions/common.State"
                        }
                    }
                }
            }
        },
        "/api/v1/movie/delete": {
            "delete": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movie"
                ],
                "summary": "删除电影(不能删除在演出计划里面的)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "x_token 用户令牌",
                        "name": "x_token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "movie_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "1001:参数有误 1003:系统错误 2001:鉴权失败",
                        "schema": {
                            "$ref": "#/definitions/common.State"
                        }
                    }
                }
            }
        },
        "/api/v1/movie/detail": {
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movie"
                ],
                "summary": "获取电影详细信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "x_token 用户令牌",
                        "name": "x_token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "movieID",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "1001:参数有误 1003:系统错误 2001:鉴权失败 ",
                        "schema": {
                            "$ref": "#/definitions/common.State"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "common.State": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "失败时返回空"
                },
                "status_code": {
                    "description": "状态码，0-成功，其他值-失败",
                    "type": "integer"
                },
                "status_msg": {
                    "description": "返回状态描述",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}